remove app directory and student directory and move files to root directory
